% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_wasa_echse.R
\name{compare_wasa_echse}
\alias{compare_wasa_echse}
\title{Comparison of simulation outputs of the WASA-SED model and the WASA engine of the ECHSE environment}
\usage{
compare_wasa_echse(resdir_echse = NULL, resdir_wasa = NULL,
  dir_out = NULL, plot_name = NULL, sub_pars = NULL,
  prec_file_wasa = NULL, resol = "hourly")
}
\value{
Function returns nothing. A plot (pdf file) will be created.
}
\description{
Function creates a plot as pdf file showing the simulated time series of the most important
water balance components.
}
\author{
Tobias Pilz \email{tpilz@uni-potsdam.de}
}
