% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydInd.R
\name{hydInd}
\alias{hydInd}
\title{Calculation of hydrological indices / streamflow signatures}
\usage{
hydInd(dat.q.ts = NULL, dat.pr.ts = NULL, na.rm = T, ignore.zeros = F,
  thresh.zero = NULL, flood.thresh = NULL)
}
\arguments{
\item{dat.q.ts}{Object of class 'xts' containing a discharge time series. Unit
must be (m3/s) whereas the resolution must be the same as for dat.pr.ts.}

\item{dat.pr.ts}{Object of class 'xts' containing a rainfall time series. Unit
must be (m3) whereas the resolution must be the same as for dat.q.ts.}

\item{na.rm}{Object of type \code{logical}. Shall \code{NA} values be removed from
\code{dat.q.ts} and \code{dat.pr.ts}? If \code{FALSE}, an error might occur for
certain calculations if \code{NA}s are found. Default: \code{TRUE}.}

\item{ignore.zeros}{Object of type \code{logical}. Shall value less than or equal
to \code{thresh.zero} be removed from \code{dat.q.ts}? Will only be applied for
certain quantile-based calculations as many small / zero values may distort the
result.}

\item{thresh.zero}{Values of \code{dat.q.ts} below this value will be treated as
zero flows (directed to \code{\link[hydrostats]{high.spells}} argument \code{ctf.threshold}).}

\item{flood.thresh}{Numeric value giving the threshold in (V/T) for the definition of
a flood event (directed to \code{\link[hydrostats]{high.spells}} argument \code{threshold}).}
}
\value{
A named vector of type numeric with the following elements:

rc: Runoff coefficient in (\%)

m_h: Average annual maximum flow in (m3/s)

m_l: Average annual minimum flow in (m3/s)

f_h: Average number of high spell events per year

f_l: Average number of low flow events per year

r_r: Average absolute daily change during rise periods within high flow events (m3/s/day)

r_f: Average absolute daily change during falling periods within high flow events (m3/s/day)

fdc_s: Average slope of flow duration curve for medium range (33% to 66%) of non-zero flows;
the higher the value the more variable the flow regime, see Sawicz et al. (2011)
}
\description{
Calculation of hydrological indices / streamflow signatures
}
\details{
Function employs the package \code{\link[hydrostats]{hydrostats}} for
certain calculations.
}
\author{
Tobias Pilz \email{tpilz@uni-potsdam.de}
}
