% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wasa_calibwrap.R
\name{wasa_calibwrap}
\alias{wasa_calibwrap}
\title{Wrapper function for calibration of the WASA-SED model}
\usage{
wasa_calibwrap(pars = NULL, wasa_app = NULL, sp_input_dir = NULL,
  meteo_dir = NULL, dir_run = paste0(tempfile(pattern = "wasa_calib_"),
  sep = "/"), sim_start = NULL, sim_end = NULL, resol = "daily",
  warmup_start = NULL, radex_file = NULL, dat_streamflow = NULL,
  dat_pr = NULL, flood_thresh = NULL, thresh_zero = NULL,
  warmup_len = 3, max_pre_runs = 20, storage_tolerance = 0.01,
  return_val = "river_flow", keep_rundir = FALSE, keep_log = FALSE,
  error2warn = FALSE)
}
\arguments{
\item{pars}{A named vector of type numeric with the values of selected parameters
(directed to \code{\link[WasaEchseTools]{wasa_modify_pars}}).}

\item{wasa_app}{Character string giving the system command of the WASA-SED application
(directed to \code{\link[WasaEchseTools]{wasa_run}}).}

\item{sp_input_dir}{Character string of the directory containing the spatial WASA-SED
input. E.g. the output of \code{\link[lumpR]{db_wasa_input}} (argument 'dest_dir')
(directed to \code{\link[WasaEchseTools]{wasa_prep_runs}}).}

\item{meteo_dir}{Character string of the directory containing the time series input
files for the WASA-SED model. Requires the files temperature.dat, radiation.dat,
humidity.dat, and rain_daily.dat or rain_hourly.dat (directed to \code{\link[WasaEchseTools]{wasa_prep_runs}}).}

\item{dir_run}{Character specifying the directory for the model run with the current
parameter realisation (directed to \code{\link[WasaEchseTools]{wasa_run}}).
Default: A temporary directory created with \code{\link{tempfile}}.}

\item{sim_start}{Object of class 'date' giving the start date of the simulation
(will be written into WASA-SED input file 'do.dat'; directed to \code{\link[WasaEchseTools]{wasa_prep_runs}}).}

\item{sim_end}{Object of class 'date' giving the end date of the simulation
(will be written into WASA-SED input file 'do.dat' directed to \code{\link[WasaEchseTools]{wasa_prep_runs}}).}

\item{resol}{Temporal resolution of the model simulations. Supported are: 'daily'
(default) and 'hourly'.}

\item{warmup_start}{An object of class 'date' giving the start date of the warm-up period.
If \code{NULL} (default), the value in do.dat (lines 4 and 6) is used.
Directed to \code{\link[WasaEchseTools]{wasa_run}}.}

\item{radex_file}{Character string of the file (including path) of preprared
extraterrestrial radiation input (named 'extraterrestrial_radiation.dat')
(directed to \code{\link[WasaEchseTools]{wasa_prep_runs}})).}

\item{dat_streamflow}{OPTIONAL: Object of class 'xts' containing a time series of
streamflow at the catchment outlet in (m3/s) in the resolution of the model run.
NA values are allowed and will be discarded for goodness of fit calculations.
Only needed if \code{return_val = 'nse'}.}

\item{dat_pr}{OPTIONAL: Object of class 'xts' containing a time series of catchment-wide
average precipitation in (m3) in the resolution of the model run. If not given (default),
it will be read (and calculated) from the WASA-SED input files if needed. However,
it is more efficient in terms of function execution time to specify it as input if needed.
Only needed if \code{return_val = 'hydInd'}.}

\item{flood_thresh}{OPTIONAL: Numeric value giving the threshold in (m3/s) for the definition of
a flood event (directed to \code{\link[WasaEchseTools]{hydInd}}). Only needed if
\code{return_val = 'hydInd'}.}

\item{thresh_zero}{OPTIONAL: Values of discharge in (m3/s) below this value will be treated as
zero flows (directed to \code{\link[WasaEchseTools]{hydInd}}). Only needed if
\code{return_val = 'hydInd'}.}

\item{warmup_len}{Integer giving the length of the warm-up period in months. Default: 3.
Directed to \code{\link[WasaEchseTools]{wasa_run}}.}

\item{max_pre_runs}{Integer specifying the maximum number of warm-up iterations to be
applied. If the relative storage change is still larger than \code{storage_tolerance}
after \code{max_pre_runs} iterations, the warm-up will be aborted and the model be run
anyway. A warning will be issued. Default: 20. Directed to \code{\link[WasaEchseTools]{wasa_run}}.}

\item{storage_tolerance}{Numeric value giving the relative change of the model's water
storages between two connsecutive warm-up runs below which the warm-up will be
concluded and the actual model simulation be started. Default: 0.01.
Directed to \code{\link[WasaEchseTools]{wasa_run}}.}

\item{return_val}{Character string specifying your choice of what this function
shall return. Default: 'river_flow'. See description of return value below.}

\item{keep_rundir}{Value of type \code{logical}. Shall directory \code{dir_run}
be retained (\code{TRUE}) or deleted (\code{FALSE}) after function execution?
Default: \code{FALSE}.}

\item{keep_log}{Value of type \code{logical}. Shall a log file of the model run be written (to \code{dir_run})?
Default: \code{FALSE}. Will be ignored if \code{keep_rundir = FALSE}.
Directed to \code{\link[WasaEchseTools]{wasa_run}}.}

\item{error2warn}{Value of type \code{logical}. Shall runtime errors of the model be
reported as a warning instead of stopping this function with an error? If so, the
model run's log file will be saved and, In case of reasonable model output, this wrapper
function will proceed as usual. If no reasonable output could be found,
a value \code{NA} will be returned. Default: \code{FALSE}.
Also directed to \code{\link[WasaEchseTools]{wasa_run}}.}
}
\value{
Function returns a vector of numeric values. Can be controlled by argument
\code{return_val}. Currently implemented are the options:

river_flow: An object of class 'xts' containing the simulated river flow leaving the
catchment outlet in m3/s for the specified simulation period and resolution.

hydInd: Named vector of hydrological indices calculated with function \code{\link[WasaEchseTools]{hydInd}}.
See function's doc for more information. This option requires the optional input arguments
\code{flood_thresh}, \code{thresh_zero}, and (optional) \code{dat_pr}.

nse: A single numeric value giving the Nash-Sutcliffe efficiency of streamflow
simulation at the catchment outlet (a common goodness of fit measure of hydrological
model runs).
}
\description{
A wrapper function, executing a simulation of the WASA-SED model for a given
parameter realisation.
}
\details{
Function is a wrapper function, internally executing functions \code{\link[WasaEchseTools]{wasa_prep_runs}},
\code{\link[WasaEchseTools]{wasa_modify_pars}}, and \code{\link[WasaEchseTools]{wasa_run}}
and processing and returning the simulation output as specified. The function can
be employed by model calibration functions such as \code{\link[HydroBayes]{dream}}
or \code{\link[ppso]{optim_dds}}, or to execute single model runs within a single
function call.
}
\author{
Tobias Pilz \email{tpilz@uni-potsdam.de}
}
