% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wasa_meteo.R
\name{wasa_meteo}
\alias{wasa_meteo}
\title{Preparation of WASA-SED time series input by IDW interpolation of climate data to subbasin centroids}
\usage{
wasa_meteo(file_target_locs = NULL, file_source_locs = NULL,
  file_result = NULL, file_ts = NULL, output_dir = NULL,
  resol = "daily", ...)
}
\arguments{
\item{file_target_locs}{Mandatory argument \code{file_targets} for
\code{\link[geostat]{externalInputLocationsTable}}. To calculate extraterrestrial
radiation, the file further needs to contain column 'lat' with the latitudes in
decimal degrees of the target locations.}

\item{file_source_locs}{Mandatory argument \code{file_sources} for
\code{\link[geostat]{externalInputLocationsTable}}.}

\item{file_result}{Mandatory argument for \code{\link[geostat]{externalInputLocationsTable}}.
File will be created within directory \code{output_dir}.}

\item{file_ts}{Named vector of file names. Required are the four files 'precip',
'temper', 'rhum', and 'glorad'. Each file is expected to hold time series of one
of the former variables for the various source locations (as named columns) given in
\code{file_source_locs}. The first column of each file, however, needs to hold a datetime
string in a format recognisable for function \code{\link[zoo]{read.zoo}}.}

\item{output_dir}{Character string of the directory for \code{file_result} and the
WASA-SED time series input files. Will be created.}

\item{resol}{Character string defining the resolution of the WASA-SED time series files.
Supported are: 'daily' (default) for daily and 'hourly' for hourly resolution. Note that
this affects only the time series file of precipitation and also defines the required resolution
of the input time series file! All other variables (temperature, humidity, radiation)
will be in daily resolution (and have to be given in daily resolution as well!).}

\item{...}{Additional arguments for \code{\link[geostat]{externalInputLocationsTable}}.
If not given, standard values will be used.}
}
\value{
Function returns nothing. WASA-SED time series files (and the output of
\code{\link[geostat]{externalInputLocationsTable}}) will be created.
}
\description{
This function performs IDW interpolation of climate data to subbasin
centroids (e.g. created with the \code{lumpR} package) using the
function \code{\link[geostat]{externalInputLocationsTable}} from the package
\code{\link{geostat}} which is available from the echse_tools repository: \url{https://github.com/echse/echse_tools}.
Extreterrestrial radiation is calculated employing function \code{\link[sirad]{extrat}}.
}
\author{
Tobias Pilz \email{tpilz@uni-potsdam.de}
}
