% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wasa_modify_pars.R
\name{wasa_modify_pars}
\alias{wasa_modify_pars}
\title{Modify input parameters of the WASA-SED model}
\usage{
wasa_modify_pars(pars = NULL, dir_run = NULL)
}
\arguments{
\item{pars}{A named vector of type numeric with the values of selected parameters.}

\item{dir_run}{Character string of the path to the WASA-SED input directory.}
}
\value{
Function returns nothing.
}
\description{
Function replaces specified parameter values within prepared WASA-SED model input
files by given substitutes.
}
\details{
IMPORTANT: Note that some parameters are factors or summands and the
parameters are updated by multiplication or summation rather than mere replacement.
Read notes below!

Parameter names below can be modified by adding prefix 'log_*'. In this case,
the value will be replaced by exp(value) internally for better handling of highly
uncertain parameters. For instance, 'kf_bedrock_f' might vary between 0.01 and 100
which can cause some numerical issues to the calibration algorithm (i.e. underrepresenting
small numbers < 1). In such a case it is better to specify, e.g. log(100) or log(0.01)
(i.e. 4.6 or -4.6, respectively) as limits for the calibration algorithm. Note that
only the natural logarithm (R function log()) is supported herein!


Currently supported are the following parameters (use these as names for the input
vector \code{pars}):

File Hillslope/vegetation.dat (equally applied to all vegetation types and node points):

 - stomr_f: Factor, somatal resistance\cr
 - rootd_f: Factor, rootdepth\cr
 - albedo_f: Factor, albedo

File Hillslope/soil.dat (equally applied to all types):

 - n_f: Factor, porosity\cr
 - cfr: Summand, added to coarse fragments\cr

File Hillslope/frac_direct_gw.dat:

 -  f_direct_gw: partitioning of groundwater into river and alluvia

File River/response.dat:

 - uhg_f: Factor, applied to both lag and response times

File Hillslope/soter.dat:

 - gw_delay_f: Factor, groundwater delay\cr
 - soildepth_f: Factor, soil depth\cr
 - kf_bedrock_f: Factor, kf of bedrock\cr
 - riverdepth_f: Factor, depth of riverbed

File Reservoir/lake_maxvol.dat:

 - f_lakemaxvol: Factor, water storage capacity for the reservoir size classes

File Others/calib_wind.dat:

 - f_wind: Factor, modification of wind speed (hard-coded in WASA-SED; calibration of evapotranspiration)

File do.dat:

 - kfcorr: Replacement of kfcorr value\cr
 - kfcorr0: Replacement of kfcorr0 value\cr
 - kfcorr_a: Replacement of kfcorr_a value\cr
 - intcf: Replacement of intcf value\cr
 - dosediment: Replacement of dosediment flag

File calibration.dat:

 - ksat_factor: Replacement of ksat factor (factor internally used in WASA-SED)
}
\author{
Tobias Pilz \email{tpilz@uni-potsdam.de}
}
