% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wasa_prep_runs.R
\name{wasa_prep_runs}
\alias{wasa_prep_runs}
\title{Preparation of WASA-SED model input}
\usage{
wasa_prep_runs(sp_input_dir = NULL, meteo_dir = NULL,
  radex_file = NULL, wasa_sim_dir = getwd(), proj_name = "wasa_run",
  sim_start = NULL, sim_end = NULL, timestep = 24, outfiles = NULL)
}
\arguments{
\item{sp_input_dir}{Character string of the directory containing the spatial WASA-SED
input. E.g. the output of \code{\link[lumpR]{db_wasa_input}} (argument 'dest_dir').}

\item{meteo_dir}{Character string of the directory containing the time series input
files for the WASA-SED model. Requires the files temperature.dat, radiation.dat,
humidity.dat, and rain_daily.dat or rain_hourly.dat.}

\item{radex_file}{Character string of the file (including path) of preprared
extraterrestrial radiation input (named 'extraterrestrial_radiation.dat').}

\item{wasa_sim_dir}{Character string of the location the output shall be written
to (directory will be created if it does not exist).}

\item{proj_name}{Character string of the current project name, i.e. the directory
within \code{wasa_sim_dir} which will contain the directories 'input' (the complete
WASA-SED input) and 'output' (WASA-SED simulation output will be written into this
directory).}

\item{sim_start}{Object of class 'date' giving the start date of the simulation
(will be written into WASA-SED input file 'do.dat').}

\item{sim_end}{Object of class 'date' giving the end date of the simulation
(will be written into WASA-SED input file 'do.dat').}

\item{timestep}{Integer giving the temporal resolution of the model run in hours.
Must be 1 (hourly) or 24 (daily resolution). Will be written into WASA-SED input
file 'do.dat'. Default: 24.}

\item{outfiles}{Character vector specifying the contents of input file 'outfiles.dat'
controlling the output to be generated by the model. See the WASA-SED socumentation
for supported values. If \code{NULL} (default), the file will not be generated and
default output will be produced.}
}
\value{
Function returns nothing.
}
\description{
Makes use of spatial model input prepared with the \code{\link[lumpR]{lumpR}} package,
pre-processed meteorological data, and further input to this function to generate a
directory for the model ready to run simulations.
}
\author{
Tobias Pilz \email{tpilz@uni-potsdam.de}
}
