% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wasa_run.R
\name{wasa_run}
\alias{wasa_run}
\title{Executes simulations with the WASA-SED model}
\usage{
wasa_run(dir_run = NULL, wasa_app = NULL, warmup_start = NULL,
  warmup_len = 3, max_pre_runs = 20, storage_tolerance = 0.01,
  log_meta = NULL, keep_log = FALSE, error2warn = FALSE)
}
\arguments{
\item{dir_run}{Character string giving the WASA-SED simulation directory. Needs
to contain sub-directories 'input' and 'output'.}

\item{wasa_app}{Character string giving the system command of the WASA-SED application.}

\item{warmup_start}{An object of class 'date' giving the start date of the warm-up period.
If \code{NULL} (default), the value in do.dat (lines 4 and 6) is used.}

\item{warmup_len}{Integer giving the length of the warm-up period in months. Default: 3.}

\item{max_pre_runs}{Integer specifying the maximum number of warm-up iterations to be
applied. If the relative storage change is still larger than \code{storage_tolerance}
after \code{max_pre_runs} iterations, the warm-up will be aborted and the model be run
anyway. A warning will be issued. Default: 20.}

\item{storage_tolerance}{Numeric value giving the relative change of the model's water
storages between two connsecutive warm-up runs below which the warm-up will be
concluded and the actual model simulation be started. Default: 0.01.}

\item{log_meta}{Character containg the name (and path) of a file into which information
about the function call will be written. See below for more information. Default:
\code{NULL}, i.e. no logile will be created. If the file already exists, the file
name to be created will be extended by a call to \code{\link{tempfile}}.}

\item{keep_log}{Value of type \code{logical}. Shall a log file of the model run be written (to \code{dir_run})?
Default: \code{FALSE}.}

\item{error2warn}{Value of type \code{logical}. Shall runtime errors of the model be
reported as a warning instead of stopping this function with an error? If so, the
model run's log file will be saved. Default: \code{FALSE}.}
}
\value{
Function returns nothing.

If argument \code{log_meta} was given, the logfile contains the following information:
\code{run_dir}: realisation of argument \code{dir_run}, \code{time_simrun}:
runtime of the simulation run (in secs.), \code{time_warmup}: total runtime for all
warm-up runs (in secs.), \code{warmup_iterations}: number of warm-up iterations,
\code{warmup_storchange}: relative storage change after the last warm-up iteration.
}
\description{
Function executes simulation with the WASA-SED model including warm-up runs to
start from stable model states as initial conditions.
}
\note{
To avoid warm-up runs, set \code{max_pre_runs} or \code{warmup_len} to zero.
}
\author{
Tobias Pilz \email{tpilz@uni-potsdam.de}
}
