% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wasa_run.R
\name{wasa_run}
\alias{wasa_run}
\title{Executes simulations with the WASA-SED model}
\usage{
wasa_run(dir_run = NULL, wasa_app = NULL, warmup_len = 3,
  max_pre_runs = 20, storage_tolerance = 0.01)
}
\arguments{
\item{dir_run}{Character string giving the WASA-SED simulation directory. Needs
to contain sub-directories 'input' and 'output'.}

\item{wasa_app}{Character string giving the system command of the WASA-SED application.}

\item{warmup_len}{Integer giving the length of the warm-up period in months. Default: 3.}

\item{max_pre_runs}{Integer specifying the maximum number of warm-up iterations to be
applied. If the relative storage change is still larger than \code{storage_tolerance}
after \code{max_pre_runs} iterations, the warm-up will be aborted and the model be run
anyway. A warning will be issued. Default: 20.}

\item{storage_tolerance}{Numeric value giving the relative change of the model's water
storages between two connsecutive warm-up runs below which the warm-up will be
concluded and the actual model simulation be started. Default: 0.01.}
}
\value{
Function returns nothing.
}
\description{
Function executes simulation with the WASA-SED model including warm-up runs to
start from stable model states as initial conditions.
}
\author{
Tobias Pilz \email{tpilz@uni-potsdam.de}
}
